Fix Pulumi Provider download URL
When PluginDownloadURL is not set, Pulumi assumes that the
plugin is downloaded from the official repository in the
Pulumi github organization. As a result, in order to install
the Chronosphere provider you need to run the
`pulumi plugin install` command with a `--server` argument
pointing it at the right place to download it.

This change simply sets the right download URL so that this
won't be necessary anymore